gameManagerにGameControllerを統合した
